Updated test suite to include all cases tested in v1.
The equality operator is constexpr in my version of the compile time string so I can assert correctness at compile time instead of having to check at run time like I do when the results are in a std::array. On the down side, I have to include the length of the expected strings, which is annoying, but I should be able to get rid of this with the static "literal_to_string" function.