Solution to compile time fizz buzz using tuples.
Converting from number to string is no longer necessary in this solution so I was able to cut the size of the file down significantly. One downside to this solution is that there is no simple ways to iterate through a tuple or write to a stream. This can be solved with more template trickery but it does complicate the answer.